feat: simplified framework with fail-closed security and a fully green gate matrix
Squash of codex/Simplified: backend/frontend simplification finalized with security hardening (fail-closed client IP resolution, Linux-safe path traversal guard, Tomcat 10.1.59 CVE bump), coverage debt settled on both stacks (lines 57.7% -> 82%, 37 backend classes + 158 frontend files cleared), docs/contracts aligned (ADR 0028, threat model T1, debt retirement policy), and the full CI gate matrix green on real runners including Testcontainers integration and Trivy supply-chain scans.

+++ 后端代码/basic-framework-boot/basic-framework-module-system/src/main/java/com/basicframework/module/system/controller/admin/notify/vo/message/NotifyMessageRespVO.java
@@ -16,7 +16,7 @@ public class NotifyMessageRespVO {
     private Long userId;
 
     @Schema(description = "用户类型，参见 UserTypeEnum 枚举", requiredMode = Schema.RequiredMode.REQUIRED, example = "1")
-    private Byte userType;
+    private Integer userType;
 
     @Schema(description = "模版编号", requiredMode = Schema.RequiredMode.REQUIRED, example = "13013")
     private Long templateId;
